requirements.txt
jupyterlab==3.6.1
pandas==1.5.3
openpyxl==3.1.2
matplotlib==3.7.1
scikit-learn==1.2.2
translate-toolkit==3.8.6

# Automatic evaluation
pdfkit==1.0.0
sacrebleu==2.3.1
prettytable==3.6.0
diff_match_patch==20200713
nltk==3.8.1
unbabel-comet==2.0.1
evaluate==0.4.0